requirements.txt
amqp==5.2.0
anyio==3.6.2
#argon2-cffi==21.3.0
#argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
billiard==4.2.0
bleach==6.0.0
blinker==1.6.3
celery==5.4.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
coverage==7.6.1
cron-descriptor==1.4.3
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==5.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.4
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.9.1
greenlet==3.0.3
huggingface-hub==0.17.2
idna==3.4
importlib_resources==6.4.4
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
kiwisolver==1.4.4
kombu==5.4.0
loguru==0.7.2
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.37
psutil==5.9.4
psycopg-binary==3.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.9.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2022.7.1
#pywin32==305
#pywinpty==2.0.10
PyYAML==6.0
#pyzmq==25.0.0
#qtconsole==5.4.0
QtPy==2.3.0
redis==5.0.8
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
safetensors==0.3.3
#scikit-learn==1.3.1
#scipy==1.11.3
#seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
#sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.31
#sqlparse==0.5.1
stack-data==0.6.2
swagger-spec-validator==3.0.4
sympy==1.12
#terminado==0.17.1
#threadpoolctl==3.1.0
#tinycss2==1.2.1
#tokenizers==0.13.3
#torch==2.0.1
#tornado==6.2
#tqdm==4.66.1
#traitlets==5.9.0
#transformers==4.33.2
#typing_extensions==4.8.0
tzdata==2024.1
#uri-template==1.2.0
#uritemplate==4.1.1
#urllib3==2.0.5
#vine==5.1.0
#wcwidth==0.2.6
#webcolors==1.12
#webencodings==0.5.1
#websocket-client==1.5.1
#Werkzeug==3.0.0
#widgetsnbextension==4.0.5

